+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/api/HoodieSink.java
@@ -1,0 +1,182 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hudi.api;
+
+import org.apache.hudi.common.model.HoodieRecord;
+import org.apache.hudi.common.model.WriteOperationType;
+import org.apache.hudi.configuration.FlinkOptions;
+import org.apache.hudi.configuration.OptionsResolver;
+import org.apache.hudi.sink.utils.Pipelines;
+import org.apache.hudi.util.ChangelogModes;
+import org.apache.hudi.util.StreamerUtil;
+
+import org.apache.flink.annotation.VisibleForTesting;
+import org.apache.flink.configuration.Configuration;
+import org.apache.flink.streaming.api.datastream.DataStream;
+import org.apache.flink.table.catalog.ResolvedSchema;
+import org.apache.flink.table.connector.ChangelogMode;
+import org.apache.flink.table.connector.sink.DataStreamSinkProvider;
+import org.apache.flink.table.connector.sink.DynamicTableSink;
+import org.apache.flink.table.types.logical.RowType;
+
+import java.util.Map;
+
+/**
+ * Hoodie api sink.
+ */
+public class HoodieSink {
+
+  private final Configuration conf;
+  private final ResolvedSchema schema;
+  private boolean overwrite = false;
+  private String operatorName;
+  private boolean isBounded;
+
+  public HoodieSink(Configuration conf, ResolvedSchema schema) {
+    this(conf, schema, false, false, "hoodie_sink");
+  }
+
+  public HoodieSink(Configuration conf, ResolvedSchema schema, boolean overwrite, boolean isBounded, String operatorName) {
+    this.conf = conf;
+    this.schema = schema;
+    this.overwrite = overwrite;
+    this.isBounded = isBounded;
+    this.operatorName = operatorName;
+  }
+
+  public DynamicTableSink.SinkRuntimeProvider getSinkRuntimeProvider(Boolean isBounded) {
+    return (DataStreamSinkProvider) dataStream -> {
+
+      // setup configuration
+      long ckpTimeout = dataStream.getExecutionEnvironment()
+          .getCheckpointConfig().getCheckpointTimeout();
+      conf.setLong(FlinkOptions.WRITE_COMMIT_ACK_TIMEOUT, ckpTimeout);
+
+      RowType rowType = (RowType) schema.toSourceRowDataType().notNull().getLogicalType();
+
+      // bulk_insert mode
+      final String writeOperation = this.conf.get(FlinkOptions.OPERATION);
+      if (WriteOperationType.fromValue(writeOperation) == WriteOperationType.BULK_INSERT) {
+        return Pipelines.bulkInsert(conf, rowType, dataStream);
+      }
+
+      // Append mode
+      if (OptionsResolver.isAppendMode(conf)) {
+        return Pipelines.append(conf, rowType, dataStream, isBounded);
+      }
+
+      // default parallelism
+      int parallelism = dataStream.getExecutionConfig().getParallelism();
+      DataStream<Object> pipeline;
+      // bootstrap
+      final DataStream<HoodieRecord> hoodieRecordDataStream =
+          Pipelines.bootstrap(conf, rowType, parallelism, dataStream, isBounded, overwrite);
+      // write pipeline
+      pipeline = Pipelines.hoodieStreamWrite(conf, parallelism, hoodieRecordDataStream);
+      // compaction
+      if (StreamerUtil.needsAsyncCompaction(conf)) {
+        return Pipelines.compact(conf, pipeline);
+      } else {
+        return Pipelines.clean(conf, pipeline);
+      }
+    };
+  }
+
+  @VisibleForTesting
+  public Configuration getConf() {
+    return this.conf;
+  }
+
+  public ChangelogMode getChangelogMode(ChangelogMode changelogMode) {
+    if (conf.getBoolean(FlinkOptions.CHANGELOG_ENABLED)) {
+      return ChangelogModes.FULL;
+    } else {
+      return ChangelogModes.UPSERT;
+    }
+  }
+
+  public HoodieSink copy() {
+    return new HoodieSink(this.conf, this.schema, this.overwrite, this.isBounded, this.operatorName);
+  }
+
+  public String asSummaryString() {
+    return operatorName;
+  }
+
+  public void applyStaticPartition(Map<String, String> partitions) {
+    // #applyOverwrite should have been invoked.
+    if (this.overwrite && partitions.size() > 0) {
+      this.conf.setString(FlinkOptions.OPERATION, WriteOperationType.INSERT_OVERWRITE.value());
+    }
+  }
+
+  public void applyOverwrite(boolean overwrite) {
+    this.overwrite = overwrite;
+    // set up the operation as INSERT_OVERWRITE_TABLE first,
+    // if there are explicit partitions, #applyStaticPartition would overwrite the option.
+    this.conf.setString(FlinkOptions.OPERATION, WriteOperationType.INSERT_OVERWRITE_TABLE.value());
+  }
+
+  public static HoodieSink.Builder builder() {
+    return new Builder();
+  }
+
+  public static class Builder {
+    private  Configuration conf;
+    private ResolvedSchema schema;
+    private boolean overwrite = false;
+    private boolean isBouned;
+    private String operatorName;
+
+    public Builder self() {
+      return this;
+    }
+
+    public Builder conf(Configuration conf) {
+      this.conf = conf;
+      return self();
+    }
+
+    public Builder schema(ResolvedSchema schema) {
+      this.schema = schema;
+      return self();
+    }
+
+    public Builder overWrite(Boolean overwrite) {
+      this.overwrite = overwrite;
+      return self();
+    }
+
+    public Builder isBouned(Boolean isBouned) {
+      this.isBouned = isBouned;
+      return self();
+    }
+
+    public Builder operatorName(String operatorName) {
+      this.operatorName = operatorName;
+      return self();
+    }
+
+    public HoodieSink build() {
+      return new HoodieSink(this.conf, this.schema, this.overwrite, this.isBouned, this.operatorName);
+    }
+
+  }
+
+}
