+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/table/HoodieTableSink.java
@@ -18,24 +18,14 @@
 
 package org.apache.hudi.table;
 
-import org.apache.hudi.common.model.HoodieRecord;
-import org.apache.hudi.common.model.WriteOperationType;
-import org.apache.hudi.configuration.FlinkOptions;
-import org.apache.hudi.configuration.OptionsResolver;
-import org.apache.hudi.sink.utils.Pipelines;
-import org.apache.hudi.util.ChangelogModes;
-import org.apache.hudi.util.StreamerUtil;
+import org.apache.hudi.api.HoodieSink;
 
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.streaming.api.datastream.DataStream;
-import org.apache.flink.table.catalog.ResolvedSchema;
 import org.apache.flink.table.connector.ChangelogMode;
-import org.apache.flink.table.connector.sink.DataStreamSinkProvider;
 import org.apache.flink.table.connector.sink.DynamicTableSink;
 import org.apache.flink.table.connector.sink.abilities.SupportsOverwrite;
 import org.apache.flink.table.connector.sink.abilities.SupportsPartitioning;
-import org.apache.flink.table.types.logical.RowType;
 
 import java.util.Map;
 
@@ -44,97 +34,44 @@ import java.util.Map;
  */
 public class HoodieTableSink implements DynamicTableSink, SupportsPartitioning, SupportsOverwrite {
 
-  private final Configuration conf;
-  private final ResolvedSchema schema;
-  private boolean overwrite = false;
+  private final HoodieSink hoodieSink;
 
-  public HoodieTableSink(Configuration conf, ResolvedSchema schema) {
-    this.conf = conf;
-    this.schema = schema;
-  }
-
-  public HoodieTableSink(Configuration conf, ResolvedSchema schema, boolean overwrite) {
-    this.conf = conf;
-    this.schema = schema;
-    this.overwrite = overwrite;
+  public HoodieTableSink(HoodieSink hoodieSink) {
+    this.hoodieSink = hoodieSink;
   }
 
   @Override
   public SinkRuntimeProvider getSinkRuntimeProvider(Context context) {
-    return (DataStreamSinkProvider) dataStream -> {
-
-      // setup configuration
-      long ckpTimeout = dataStream.getExecutionEnvironment()
-          .getCheckpointConfig().getCheckpointTimeout();
-      conf.setLong(FlinkOptions.WRITE_COMMIT_ACK_TIMEOUT, ckpTimeout);
-
-      RowType rowType = (RowType) schema.toSourceRowDataType().notNull().getLogicalType();
-
-      // bulk_insert mode
-      final String writeOperation = this.conf.get(FlinkOptions.OPERATION);
-      if (WriteOperationType.fromValue(writeOperation) == WriteOperationType.BULK_INSERT) {
-        return Pipelines.bulkInsert(conf, rowType, dataStream);
-      }
-
-      // Append mode
-      if (OptionsResolver.isAppendMode(conf)) {
-        return Pipelines.append(conf, rowType, dataStream, context.isBounded());
-      }
-
-      // default parallelism
-      int parallelism = dataStream.getExecutionConfig().getParallelism();
-      DataStream<Object> pipeline;
-      // bootstrap
-      final DataStream<HoodieRecord> hoodieRecordDataStream =
-          Pipelines.bootstrap(conf, rowType, parallelism, dataStream, context.isBounded(), overwrite);
-      // write pipeline
-      pipeline = Pipelines.hoodieStreamWrite(conf, parallelism, hoodieRecordDataStream);
-      // compaction
-      if (StreamerUtil.needsAsyncCompaction(conf)) {
-        return Pipelines.compact(conf, pipeline);
-      } else {
-        return Pipelines.clean(conf, pipeline);
-      }
-    };
+    return hoodieSink.getSinkRuntimeProvider(context.isBounded());
   }
 
   @VisibleForTesting
   public Configuration getConf() {
-    return this.conf;
+    return hoodieSink.getConf();
   }
 
   @Override
   public ChangelogMode getChangelogMode(ChangelogMode changelogMode) {
-    if (conf.getBoolean(FlinkOptions.CHANGELOG_ENABLED)) {
-      return ChangelogModes.FULL;
-    } else {
-      return ChangelogModes.UPSERT;
-    }
+    return hoodieSink.getChangelogMode(changelogMode);
   }
 
   @Override
   public DynamicTableSink copy() {
-    return new HoodieTableSink(this.conf, this.schema, this.overwrite);
+    return new HoodieTableSink(this.hoodieSink);
   }
 
   @Override
   public String asSummaryString() {
-    return "HoodieTableSink";
+    return hoodieSink.asSummaryString();
   }
 
   @Override
   public void applyStaticPartition(Map<String, String> partitions) {
-    // #applyOverwrite should have been invoked.
-    if (this.overwrite && partitions.size() > 0) {
-      this.conf.setString(FlinkOptions.OPERATION, WriteOperationType.INSERT_OVERWRITE.value());
-    }
+    hoodieSink.applyStaticPartition(partitions);
   }
 
   @Override
   public void applyOverwrite(boolean overwrite) {
-    this.overwrite = overwrite;
-    // set up the operation as INSERT_OVERWRITE_TABLE first,
-    // if there are explicit partitions, #applyStaticPartition would overwrite the option.
-    this.conf.setString(FlinkOptions.OPERATION, WriteOperationType.INSERT_OVERWRITE_TABLE.value());
+    hoodieSink.applyOverwrite(overwrite);
   }
 }
